fix(frontent): component refactor and packages updated
