+++ front-syloper/package.json
@@ -20,11 +20,10 @@
     "@fontsource/work-sans": "^4.5.2",
     "babel-plugin-styled-components": "^2.0.2",
     "gatsby": "^4.2.0",
-    "gatsby-plugin-sass": "^5.3.0",
     "gatsby-plugin-styled-components": "^5.3.0",
+    "gatsby-plugin-svgr": "^3.0.0-beta.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "sass": "^1.44.0",
     "styled-components": "^5.3.3",
     "styled-reset": "^4.3.4"
   },
@@ -36,6 +35,12 @@
     "@storybook/builder-webpack5": "^6.3.12",
     "@storybook/manager-webpack5": "^6.3.12",
     "@storybook/react": "^6.3.12",
-    "babel-loader": "^8.2.3"
+    "babel-loader": "^8.2.3",
+    "eslint": "^8.4.1",
+    "eslint-config-airbnb": "^18.2.1",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-react": "^7.27.1",
+    "eslint-plugin-react-hooks": "^4.3.0"
   }
 }
