+++ front-syloper/src/components/Layout/index.js
@@ -1,13 +1,14 @@
 import React from 'react';
-import { GlobalStyle } from '../../theme/globalStyle'
+import { ThemeProvider } from 'styled-components';
+import Theme, { GlobalStyle } from '../../theme';
 
 const Layout = ({ children }) => {
-    return ( 
-        <>
-            <GlobalStyle />
-            {children}
-        </>
-     );
+  return (
+    <ThemeProvider theme={Theme}>
+      <GlobalStyle />
+      {children}
+    </ThemeProvider>
+  );
 }
- 
+
 export default Layout;
