+++ front-syloper/src/components/ui/Input/index.js
@@ -1,0 +1,3 @@
+import InputStyled from './styles/InputStyled'
+
+export default InputStyled;
