+++ front-syloper/src/components/ui/Input/styles/InputStyled.js
@@ -1,13 +1,12 @@
-import React from 'react';
 import styled from 'styled-components';
+import PropTypes from 'prop-types';
 
-const InputComp = styled.input`
+const InputStyled = styled.input`
     width: 100%;
     border: none;
     box-sizing: border-box;
     padding: 14px 16px;
     background-color: #fff;
-    margin: 0 0 20px 0;
     font-size: 1.15rem;
     color: ${({ theme }) => theme.color.text };
     margin: ${props => props.margin};
@@ -17,21 +16,14 @@ const InputComp = styled.input`
     @media only screen and (max-width: 768px){
         font-size: 1.1rem;
     }
-    
 `;
 
-const Input = ({ type, placeholder, onChange, value, name, className, margin }) => {
-    return ( 
-        <InputComp
-            type={type}
-            placeholder={placeholder}
-            onChange={onChange}
-            value={value}
-            name={name}
-            className={className}
-            margin={margin}
-        ></InputComp>
-     );
+InputStyled.propTypes = {
+  margin: PropTypes.string.isRequired,
 }
- 
-export default Input;
+
+InputStyled.defaultProps = {
+  margin: '0 0 20px 0',
+}
+
+export default InputStyled;
