--- front-syloper/src/components/ui/SyloperLogo/index.jsx
@@ -1,42 +1,0 @@
-import React from 'React';
-
-const SyloperLogo = () => {
-    return(
-        <svg x="0px" y="0px" viewBox="0 0 487.5 119.3" width="100%">
-            <g>
-                <g>
-                    <path fill="#2A3235" className="st0" d="M183.1,70.3c-0.1-1.3,0.7-2,1.9-2h8.3c1.1,0,1.7,0.6,2,1.8c0.6,2.7,2.9,4.1,7.3,4.1c4.9,0,6.5-1.9,6.5-4.1
-                        c0-3.1-3.8-4.5-8.4-6c-7.5-2.6-16.9-5.7-16.9-17.3c0-10.6,8.6-15.9,17.6-15.9c8.9,0,17.3,4.8,18.2,15.7c0.1,1.3-0.7,2-1.9,2h-8.3
-                        c-1.1,0-1.7-0.6-2-1.8c-0.6-2.8-2.6-4.1-6.1-4.1c-3.8,0-5.5,1.9-5.5,4.1c0,3.2,3.1,4.7,7.5,6.3c7.6,2.6,17.9,5.4,17.9,17
-                        c0,10.6-8.5,15.9-18.6,15.9C192.6,86.1,183.9,81.3,183.1,70.3z"/>
-                    <path fill="#2A3235" className="st0" d="M265.5,34.3l-15.7,27.5v21.3c0,1.2-0.7,1.9-1.9,1.9h-8.3c-1.2,0-1.9-0.7-1.9-1.9V61.8L222,34.3
-                        c-0.8-1.4-0.2-2.3,1.3-2.3h9.7c1.1,0,1.7,0.5,2.2,1.4l8.6,16.5l8.6-16.5c0.5-0.9,1.1-1.4,2.2-1.4h9.7
-                        C265.5,32,266.1,32.9,265.5,34.3z"/>
-                    <path fill="#2A3235" className="st0" d="M300.8,75.2v7.9c0,1.2-0.7,1.9-1.9,1.9h-22.3h-4.2c-1.2,0-1.9-0.7-1.9-1.9V33.9c0-1.2,0.7-1.9,1.9-1.9h8.3
-                        c1.2,0,1.9,0.7,1.9,1.9v39.4h16.3C300,73.3,300.8,74,300.8,75.2z"/>
-                    <path fill="#2A3235" className="st0" d="M301.5,58.5c0-15.6,12.3-27.5,27.5-27.5s27.5,12,27.5,27.5c0,15.6-12.3,27.5-27.5,27.5
-                        C313.9,86.1,301.5,74.1,301.5,58.5z M344.5,58.5c0-9.5-6.8-15.7-15.4-15.7s-15.4,6.3-15.4,15.7c0,9.5,6.8,15.7,15.4,15.7
-                        C337.6,74.3,344.5,68,344.5,58.5z"/>
-                    <path fill="#2A3235" className="st0" d="M402,50.2c0,10.1-8.1,18.2-18.5,18.2h-7.9v14.7c0,1.2-0.7,1.9-1.9,1.9h-8.3c-1.2,0-1.9-0.7-1.9-1.9V33.9
-                        c0-1.2,0.7-1.9,1.9-1.9h10.2h7.9C394,32,402,40.1,402,50.2z M390,50.2c0-3.9-2.7-6.8-6.4-6.8h-7.9V57h7.9
-                        C387.1,57,390,54.1,390,50.2z"/>
-                    <path fill="#2A3235" className="st0" d="M437.5,75.2v7.9c0,1.2-0.7,1.9-1.9,1.9h-26.1h-1.9c-1.2,0-1.9-0.7-1.9-1.9V33.9c0-1.2,0.7-1.9,1.9-1.9h10.2
-                        h17.4c1.2,0,1.9,0.7,1.9,1.9v7.9c0,1.2-0.7,1.9-1.9,1.9h-17.4v8.9h15.5c1.2,0,1.9,0.7,1.9,1.9v7.6c0,1.2-0.7,1.9-1.9,1.9h-15.5
-                        v9.5h17.8C436.8,73.3,437.5,74,437.5,75.2z"/>
-                    <path fill="#2A3235" className="st0" d="M487,82.7c1,1.2,0.4,2.3-1.1,2.3h-9.7c-1,0-1.7-0.3-2.3-1.1l-12.9-16.3h-4.7v15.5c0,1.2-0.7,1.9-1.9,1.9h-8.3
-                        c-1.2,0-1.9-0.7-1.9-1.9V33.9c0-1.2,0.7-1.9,1.9-1.9h10.2h8.3c10.1,0,18.2,8.1,18.2,18.2c0,6.3-3.8,12.1-9.4,15.2L487,82.7
-                        L487,82.7z M456.4,57h8.3c3.3,0,6-2.9,6-6.8s-2.7-6.8-6-6.8h-8.3V57z"/>
-                </g>
-                <g>
-                    <path fill="#00B96B" className="st1" d="M158.4,54.3L106.6,2.5c-1.5-1.5-3.4-2.3-5.5-2.3s-4,0.8-5.5,2.3L80.2,17.9L69.8,28.3L43.7,54.4
-                        c-3,3-3,7.9,0,10.9l26.1,26.1L80.2,81l-21-21.3l21.1-21.2l10.4-10.4l10.4-10.4l42,42l-42,42L90.7,91.3l-10.4,10.4L95.6,117
-                        c1.4,1.4,3.4,2.3,5.5,2.3c2,0,4-0.8,5.5-2.3l51.8-51.8C161.4,62.2,161.4,57.3,158.4,54.3z"/>
-                    <path fill="#2A3235" className="st0" d="M116.8,54.3L90.8,28.2L80.3,38.6l21.1,21.2L80.3,80.9L69.9,91.3l-10.4,10.4l-42-42l42-42l10.4,10.4l10.4-10.4
-                        L64.9,2.3C63.4,0.8,61.5,0,59.4,0s-4,0.8-5.5,2.3l-51.7,52c-3,3-3,7.9,0,10.9L54.1,117c1.4,1.4,3.4,2.3,5.5,2.3c2,0,4-0.8,5.5-2.3
-                        l15.4-15.3l10.4-10.4L117,65.2C119.9,62.2,119.9,57.3,116.8,54.3z"/>
-                </g>
-            </g>
-        </svg>
-    )
-}
-export default SyloperLogo;
