+++ front-syloper/src/images/index.js
@@ -1,0 +1,5 @@
+import { ReactComponent as SyloperLogo } from './syloper-logo.svg';
+
+export {
+  SyloperLogo,
+}
