+++ front-syloper/src/pages/404.js
@@ -1,5 +1,6 @@
 import * as React from "react"
 import { Link } from "gatsby"
+// REFACTORIZAR
 
 // styles
 const pageStyles = {
