+++ front-syloper/src/pages/recover-password.js
@@ -3,6 +3,7 @@ import Layout from "../components/layout";
 import styled from "styled-components"
 import RecoverPassword from "../components/Login/RecoverPassword";
 
+// REFACTORIZAR
 
 const MainLogin = styled.div`
   width: 100%;
